requirements.txt
Django==4.2.7
Pillow==10.4.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.0.0